+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.3.9 - 2017-05-02
 
 ### Added
-- gaussLuSolver
+- gaussLuSolver, gaussPivotingSolver
 
 ## 0.3.8 - 2017-04-14